Final Update For Now
I will be going into the primal code and doing a lot of debugging. I will be testing when the time comes. And in the mean time I will start doing an emergency shutdown branch, and some sub branches. Then  I will take it to RedSnippe And convert it. And I will do BotGCM to debug so yeah.